Move user commands to `wcf\command\user`
