+++ wcfsetup/install/files/lib/command/user/DisableAvatar.class.php
@@ -1,34 +1,37 @@
 <?php
 
-namespace wcf\system\user\command;
+namespace wcf\command\user;
 
 use wcf\data\user\User;
 use wcf\data\user\UserEditor;
-use wcf\event\user\UserCoverPhotoEnabled;
+use wcf\event\user\UserAvatarDisabled;
 use wcf\system\event\EventHandler;
 
 /**
- * Enable a user's cover photo.
+ * Disable a user's avatar.
  *
  * @author Olaf Braun
  * @copyright 2001-2025 WoltLab GmbH
  * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @since 6.3
  */
-final class EnableCoverPhoto
+final class DisableAvatar
 {
     public function __construct(
         private readonly User $user,
-    ) {
-    }
+        private readonly string $reason,
+        private readonly int $expires = 0,
+    ) {}
 
     public function __invoke(): void
     {
         (new UserEditor($this->user))->update([
-            'disableCoverPhoto' => 0,
+            'disableAvatar' => 1,
+            'disableAvatarReason' => $this->reason,
+            'disableAvatarExpires' => $this->expires,
         ]);
 
-        $event = new UserCoverPhotoEnabled($this->user);
+        $event = new UserAvatarDisabled($this->user);
         EventHandler::getInstance()->fire($event);
     }
 }
