+++ wcfsetup/install/files/lib/command/user/DisableCoverPhoto.class.php
@@ -1,38 +1,37 @@
 <?php
 
-namespace wcf\system\user\command;
+namespace wcf\command\user;
 
 use wcf\data\user\User;
 use wcf\data\user\UserEditor;
-use wcf\event\user\UserAvatarDisabled;
+use wcf\event\user\UserCoverPhotoDisabled;
 use wcf\system\event\EventHandler;
 
 /**
- * Disable a user's avatar.
+ * Disable a user's cover photo.
  *
  * @author Olaf Braun
  * @copyright 2001-2025 WoltLab GmbH
  * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @since 6.3
  */
-final class DisableAvatar
+final class DisableCoverPhoto
 {
     public function __construct(
         private readonly User $user,
         private readonly string $reason,
         private readonly int $expires = 0,
-    ) {
-    }
+    ) {}
 
     public function __invoke(): void
     {
         (new UserEditor($this->user))->update([
-            'disableAvatar' => 1,
-            'disableAvatarReason' => $this->reason,
-            'disableAvatarExpires' => $this->expires,
+            'disableCoverPhoto' => 1,
+            'disableCoverPhotoReason' => $this->reason,
+            'disableCoverPhotoExpires' => $this->expires,
         ]);
 
-        $event = new UserAvatarDisabled($this->user);
+        $event = new UserCoverPhotoDisabled($this->user);
         EventHandler::getInstance()->fire($event);
     }
 }
