+++ wcfsetup/install/files/lib/command/user/EnableAvatar.class.php
@@ -1,38 +1,33 @@
 <?php
 
-namespace wcf\system\user\command;
+namespace wcf\command\user;
 
 use wcf\data\user\User;
 use wcf\data\user\UserEditor;
-use wcf\event\user\UserCoverPhotoDisabled;
+use wcf\event\user\UserAvatarEnabled;
 use wcf\system\event\EventHandler;
 
 /**
- * Disable a user's cover photo.
+ * Enable a user's avatar.
  *
  * @author Olaf Braun
  * @copyright 2001-2025 WoltLab GmbH
  * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @since 6.3
  */
-final class DisableCoverPhoto
+final class EnableAvatar
 {
     public function __construct(
         private readonly User $user,
-        private readonly string $reason,
-        private readonly int $expires = 0,
-    ) {
-    }
+    ) {}
 
     public function __invoke(): void
     {
         (new UserEditor($this->user))->update([
-            'disableCoverPhoto' => 1,
-            'disableCoverPhotoReason' => $this->reason,
-            'disableCoverPhotoExpires' => $this->expires,
+            'disableAvatar' => 0,
         ]);
 
-        $event = new UserCoverPhotoDisabled($this->user);
+        $event = new UserAvatarEnabled($this->user);
         EventHandler::getInstance()->fire($event);
     }
 }
